+++ packages/ocean-react/src/CardListAction/__tests__/CardListAction.test.tsx
@@ -4,7 +4,7 @@ import { PlaceholderOutline, Trash, Pencil } from '@useblu/ocean-icons-react';
 import CardListAction from '../CardListAction';
 import Badge from '../../Badge';
 import Tag from '../../Tag';
-import type { ActionItem } from '../../InternalListActions';
+import type { ActionItem } from '../../_shared/components/InternalListActions';
 
 describe('CardListAction', () => {
   describe('Basic Rendering', () => {